+++ sys/src/libc/fmt/vfprint.c
@@ -12,7 +12,7 @@ _fmtFdFlush(Fmt *f)
 	int n;
 
 	n = (char*)f->to - (char*)f->start;
-	if(n && write((int)f->farg, f->start, n) != n)
+	if(n && write((int)(uintptr)f->farg, f->start, n) != n)
 		return 0;
 	f->to = f->start;
 	return 1;
